refactor: 957 - get rid of unused data files
