--- dev-data/communication.data.ts
@@ -1,11 +1,0 @@
-import { DISCORD_LINKS } from './course-titles.data';
-
-export const RS_DOCS_COMMUNICATION_LINK = 'https://docs.rs.school/#/rs-school-chats';
-export const RS_DOCS_EN_LINK = 'https://docs.rs.school/#/en/';
-export const RS_DOCS_TELEGRAM_CHATS_LINK = 'https://docs.rs.school/#/rs-school-chats?id=telegram';
-export const JS_EN_TELEGRAM_CHAT_LINK = 'https://t.me/RSS_EN';
-export const MENTOR_ONBOARD_TELEGRAM_EN = 'https://t.me/+Q6gN_sFeedEwMmEy';
-export const MENTOR_ONBOARD_TELEGRAM_RU = 'https://t.me/+Dhj-ATXo5KoyN2My';
-export const ANNOUNCEMENT_TELEGRAM_LINK = 'https://t.me/AfishaRSSchool';
-
-export type CourseNamesChannels = keyof typeof DISCORD_LINKS;
