+++ dev-data/index.ts
@@ -1,22 +1,3 @@
-export type { LinkList } from './study-path-data.types';
-
-export {
-  ANNOUNCEMENT_TELEGRAM_LINK,
-  type CourseNamesChannels,
-  JS_EN_TELEGRAM_CHAT_LINK,
-  MENTOR_ONBOARD_TELEGRAM_EN,
-  MENTOR_ONBOARD_TELEGRAM_RU,
-  RS_DOCS_COMMUNICATION_LINK,
-  RS_DOCS_EN_LINK,
-  RS_DOCS_TELEGRAM_CHATS_LINK,
-} from './communication.data';
-
-export {
-  COURSE_TITLES,
-  type CourseNames,
-  type CourseNamesKeys,
-  DISCORD_LINKS,
-} from './course-titles.data';
 export { communityMenuStaticLinks, schoolMenuStaticLinks } from './school-menu-links';
 export { courseStatus, heroCourseLocalized } from './hero-course.data';
 export { donateOptions } from './donate-options.data';
