--- dev-data/study-path-data.types.ts
@@ -1,9 +1,0 @@
-type ItemWithLink = {
-  id: number;
-  text: string;
-  title: string;
-  link: string;
-  external?: boolean;
-};
-
-export type LinkList = ItemWithLink[];
