+++ src/entities/course/types.ts
@@ -3,8 +3,8 @@ import { StaticImageData } from 'next/image';
 import { API_COURSES_IDS_DICTIONARY } from '@/entities/course/constants';
 import { COURSE_LINKS } from '@/shared/constants';
 import { ApiResourceLocale, Language, TypeCourseSkeleton } from '@/shared/types';
+import { CourseNamesKeys } from '@/shared/types/types';
 import type { EntryCollection } from 'contentful';
-import { CourseNamesKeys } from 'data';
 
 type CourseLinks = typeof COURSE_LINKS;
