+++ src/entities/course/ui/course-card/course-card.test.tsx
@@ -4,7 +4,7 @@ import { beforeEach, describe, expect, it } from 'vitest';
 import { CourseCard } from './course-card';
 import { mockedCourses } from '@/shared/__tests__/constants';
 import { renderWithRouter } from '@/shared/__tests__/utils';
-import { COURSE_TITLES } from 'data';
+import { COURSE_TITLES } from '@/shared/constants';
 
 describe('CourseCard', () => {
   describe.each(mockedCourses)('Testing course card $title', (course) => {
