+++ src/entities/mentor/index.ts
@@ -1,4 +1,3 @@
-export type { MentorFeedback } from './types';
 export { MentorApi } from './api/mentor-api';
 export { MentorFeedbackCard } from './ui/mentor-feedback-card/mentor-feedback-card';
 export { mentorStore } from './model/store';
