+++ src/shared/constants.ts
@@ -144,3 +144,16 @@ export const WIDGET_TITLE_MOD_MAP = new Map<number | undefined, WidgetTitleVaria
   [0, 'asterisk'],
   [1, 'lines'],
 ]);
+
+export const COURSE_TITLES = {
+  JS_PRESCHOOL_RU: 'JS / Front-end Pre-school RU',
+  JS_EN: 'JS / Front-end EN',
+  JS_RU: 'JS / Front-end RU',
+  REACT: 'React',
+  ANGULAR: 'Angular',
+  NODE: 'Node.js',
+  AWS_FUNDAMENTALS: 'AWS Fundamentals',
+  AWS_CLOUD_DEVELOPER: 'AWS Cloud Developer',
+  AWS_DEVOPS: 'AWS DevOps',
+  AWS_AI: 'AWS AI',
+} as const;
