+++ src/shared/helpers/get-actual-data.test.ts
@@ -5,10 +5,9 @@ import { MOCKED_IMAGE_PATH } from '../__tests__/constants';
 import { Course } from '@/entities/course';
 import { COURSE_DATE_FORMAT } from '@/entities/course/constants';
 import { Event } from '@/entities/event';
-import { COURSE_LINKS, TO_BE_DETERMINED } from '@/shared/constants';
+import { COURSE_LINKS, COURSE_TITLES, TO_BE_DETERMINED } from '@/shared/constants';
 import { dayJS } from '@/shared/helpers/day-js';
 import { isCourse } from '@/shared/helpers/is-course';
-import { COURSE_TITLES } from 'data';
 
 const staleAfterDays = 14;
