--- src/shared/helpers/get-course-language.ts
@@ -1,9 +1,0 @@
-import { selectCourse } from '../hooks/use-course-by-title/utils/select-course';
-import { Language } from '../types';
-import { CourseNamesKeys } from 'data';
-
-export async function getCourseLanguage(courseName: CourseNamesKeys): Promise<Language> {
-  const course = await selectCourse(courseName);
-
-  return course.language;
-}
