--- src/shared/ui/text-with-link/index.ts
@@ -1,1 +1,0 @@
-export { TextWithLink } from './text-with-link';
